Fix support for /etc/aliases
Prior to this commit we used postmap to create the database for all
files, including /etc/aliases. However, the traditional sendmail format
of /etc/aliases differs from the expected input for postmap. The
sendmail format uses a colon and whitespace to separate keys and values,
whereas postfix's lookup format, just uses whitespace. This necessitates
using postalias when creating databases for sendmail alias files.

I'm not sure how the postfix acceptance test passed prior to this
change, but at least in my testing this change is required for it to
pass.